+++ docs/en/changes/changes.md
@@ -94,6 +94,7 @@
 * Fix otlp log handler reponse error and otlp span convert error.
 * Fix service_relation source layer in mq entry span analyse.
 * Fix metrics comparison in promql with bool modifier.
+* Add rate limiter for Zipkin trace receiver to limit maximum spans per second.
 * Open `health-checker` module by default due to latest UI changes. Change the default check period to 30s.
 
 #### UI